requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
boto3==1.34.0
botocore==1.34.16
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==3.3.6
django-push-notifications==3.0.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
httpagentparser==1.9.5
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
kanpai==0.1.15
MarkupPy==1.14
msgpack==1.0.7
numpy==1.26.3
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
s3transfer==0.9.0
service-identity==23.1.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.1
